Address review suggestions
